Reconnection mechanism for events and records #20
