+++ src/utils/listener.js
@@ -1,4 +1,5 @@
 var C = require( '../constants/constants' );
+var ReconnectionNotifier = require( './reconnection-notifier' );
 
 var Listener = function( type, pattern, callback, options, client, connection ) {
     this._type = type;
@@ -8,11 +9,13 @@ var Listener = function( type, pattern, callback, options, client, connection ) 
     this._client = client;
     this._connection = connection;
     this._ackTimeout = setTimeout( this._onAckTimeout.bind( this ), this._options.subscriptionTimeout );
-    this._connection.sendMsg( this._type, C.ACTIONS.LISTEN, [ this._pattern ] );
+    this._reconnectionNotifier = new ReconnectionNotifier( client, this._sendListen.bind( this ) );
+    this._sendListen();
 };
 
 Listener.prototype.destroy = function() {
     this._connection.sendMsg( this._type, C.ACTIONS.UNLISTEN, [ this._pattern ] );
+    this._reconnectionNotifier.destroy();
     this._callback = null;
     this._pattern = null;
     this._client = null;
@@ -26,6 +29,10 @@ Listener.prototype._$onMessage = function( message ) {
         var isFound = message.action === C.ACTIONS.SUBSCRIPTION_FOR_PATTERN_FOUND;
         this._callback( message.data[ 1 ], isFound );
     }
+};
+
+Listener.prototype._sendListen = function() {
+    this._connection.sendMsg( this._type, C.ACTIONS.LISTEN, [ this._pattern ] );   
 };
 
 Listener.prototype._onAckTimeout = function() {
